Updated with future implementation code
In the near future, the Vicon client implementation will become an optional or infrequent requirement. In addition an implementation for a standard for controlling robots through CRPI will be implemented so as to avoid future confusions with other versions of this code. This will allow other Unity or standalone projects to simply send a command string.